refactoring scrapers dir
