+++ scraper/parsers/parseOwners.js
